move promqlengine package
